added css shopping-list
