+++ src/sass/components/_shopping-list.scss
@@ -17,6 +17,10 @@
     letter-spacing: -0.04em;
 
     color: var(--main-text-cl-light-theme);
+
+    html.dark & {
+        color: var(--main-text-cl-dark-theme);
+    }
 }
 
 .shopping-list__LIST {
@@ -35,10 +39,17 @@
 
     border: 2px solid rgba(79, 46, 232, 0.4);
     border-radius: 16px;
+    @media screen and (min-width: 768px) {
+            padding: 24px;
+        }
 }
 
 .shopping-list__description{
     margin-top: 20px;
+
+    html.dark & {
+        color: var(--main-text-cl-dark-theme);
+    }
 }
 
 .shopping-list__general-information{
@@ -46,9 +57,16 @@
 }
 
 .shopping-list__img {
-    margin-right: 20px;
     min-width: 98px;
     height: 139px;
+    margin-right: 20px;
+    @media screen and (min-width: 768px) {
+        min-width: 116px;
+        height: 165px;
+        margin-right: 24px;
+    }
+
+
     overflow: hidden;
     border-radius: 8px;
 }
@@ -63,6 +81,28 @@
 
     letter-spacing: -0.04em;
     color: var(--secondary-text-cl);
+    
+    @media screen and (min-width: 768px) {
+            display: none;
+        }
+
+    &--near {
+        display: none;
+        @media screen and (min-width: 768px) {       
+            display: block;
+            position: absolute;;
+            bottom: 24px;
+            
+            font-style: italic;
+            font-weight: 400;
+            font-size: 12px;
+            line-height: 14px;
+            /* identical to box height, or 117% */
+        
+            letter-spacing: -0.04em;
+            color: var(--secondary-text-cl);
+        }
+    }
 }
 
 .shopping-list__title-book{
@@ -74,10 +114,14 @@
     text-transform: uppercase;
 
     color: var(--main-text-cl-light-theme);
+    html.dark & {
+            color: var(--main-text-cl-dark-theme);
+        }
 }
 
 .shopping-list__list-name {
     font-weight: 400;
+    margin-top: 4px;
     font-size: 12px;
     line-height: 14px;
     /* identical to box height, or 117% */
@@ -89,15 +133,82 @@
 
 .shopping-list__icon-deleteBtn {
     position: absolute;
-    outline: 1px solid tomato;
     top: 14px;
     right: 14px;
+
+    overflow: hidden;
+    border-radius: 50%;
+    cursor: pointer;
+
+    @media screen and (min-width: 768px) {
+        top: 24px;
+        right: 24px;
+    }
+
+    &:hover,
+    &:focus {
+        box-shadow: 0px 3px 3px rgba(0, 0, 0, 0.15);
+        scale: var(--transform-scale);
+        transition: scale var(--timing-function), box-shadow var(--timing-function);
+    }
 }
 
 .shopping-list__shop-list {
-    // display: flex;
+    display: flex;
+    align-items: center;
+    max-width: 136px;
+    max-height: 30px;
     gap: 14px;
     margin-top: 24px;
+    
+    @media screen and (min-width: 768px) {
+        position: absolute;
+            bottom: 24px;
+            right: 24px;
+            gap: 16px;
+        }
+    
+}
+
+.shopping-list__shop-lis--img {
+    filter: grayscale(100%);
+    mix-blend-mode: luminosity;
+    html.dark &{
+        filter: grayscale(0%) brightness(1);
+        width: 28px;
+        &:hover,
+        &:focus {
+            transition: scale var(--timing-function);
+        }
+    }
+    html.dark &.logoAmazon {
+        filter: brightness(0) invert(1);
+        width: 48px;
+    
+        &:hover,
+        &:focus {
+            transition: scale var(--timing-function);
+        }
+    }
+        html.dark li &.bookShop {
+            width: 32px;
+            filter: grayscale(0%) brightness(1) invert(0);
+    
+            &:hover,
+            &:focus {
+                transition: scale var(--timing-function);
+            }
+        }
+
+    &:hover,
+    &:focus {
+        scale: var(--transform-scale);
+        filter: grayscale(0%) drop-shadow(0px 3px 3px rgba(0, 0, 0, 0.15));
+        transition: filter var(--timing-function), scale var(--timing-function);
+    }
+
+
+    // mix-blend-mode: luminosity;
 }
 
 .shopping-list__description{
@@ -108,6 +219,33 @@
 
     letter-spacing: -0.04em;
     color: var(--main-text-cl-light-theme);
+
+    html.dark & {
+            color: var(--main-text-cl-dark-theme);
+        }
+
+    @media screen and (min-width: 768px) {
+            display: none;
+        }
+
+    &--near {
+            display: none;
+            @media screen and (min-width: 768px) {
+                display: block;
+                margin-top: 14px;
+                font-weight: 400;
+                font-size: 14px;
+                line-height: 18px;
+                /* or 129% */
+            
+                letter-spacing: -0.04em;
+                color: var(--main-text-cl-light-theme);
+            
+                html.dark & {
+                    color: var(--main-text-cl-dark-theme);
+                }
+            }
+        }
 }
 
 .shopping-list__clear{
