+++ src/sass/index.scss
@@ -16,3 +16,4 @@
 @import './components/hero-category';
 @import './components/modal-window-auth';
 @import './components/shopping-list';
+@import './components/support-ukraine'
