+++ src/sass/utils/_variables.scss
@@ -10,5 +10,6 @@
 
 
   //  === Other ===
+  --transform-scale: 1.015;
   --timing-function: 250ms cubic-bezier(0.4, 0, 0.2, 1);
 }
