+++ src/shopping-list.html
@@ -9,6 +9,7 @@
   </head>
   <body>
     <!-- ❗️❗️❗️ Add partials like so for every page -->
+    <include src="./toggle-color-theme.html"></include>
     <include src="./partials/_shoppingList.html"></include>
     <include src="./partials/support-ukraine.html"></include>
   </body>
