+++ src/shopping-list.js
@@ -1,2 +1,3 @@
 import './js/support-ukraine';
-import './js/shopping-list';
+import './js/shopping_list/shopping-list';
+import './js/toggle-color-theme';
